azure_text2sql/sql_agent_project/requirements.txt
pyodbc
python-dotenv
langchain
langchain-openai
langgraph
pyyaml